# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: BookSlot ERP - Multi-tenant Appointment Booking Platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /businesses:
    get:
      operationId: listBusinesses
      tags: [businesses]
      summary: List all active businesses (public)
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: city
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Business"

    post:
      operationId: createBusiness
      tags: [businesses]
      summary: Register a new business
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBusinessBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Business"
...
  /businesses/{businessId}:
    get:
      operationId: getBusiness
      tags: [businesses]
      summary: Get business by ID (public)
      parameters:
        - name: businessId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Business"

    put:
      operationId: updateBusiness
      tags: [businesses]
      summary: Update business profile
      parameters:
        - name: businessId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBusinessBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Business"
  /businesses/{businessId}/members:
    get:
      operationId: listBusinessMembers
      tags: [businesses]
      summary: List staff members of a business
      parameters:
        - name: businessId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BusinessMember"

    post:
      operationId: addBusinessMember
      tags: [businesses]
      summary: Add a staff member to the business
      parameters:
        - name: businessId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddBusinessMemberBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BusinessMember"
  /businesses/{businessId}/members/{memberId}:
    delete:
      operationId: removeBusinessMember
      tags: [businesses]
      summary: Remove a staff member
      parameters:
        - name: businessId
          in: path
          required: true
          schema:
            type: integer
        - name: memberId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed
  /businesses/{businessId}/stats:
    get:
      operationId: getBusinessStats
      tags: [businesses]
      summary: Business-level stats (bookings, appointments, revenue)
      parameters:
        - name: businessId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BusinessStats"
  /businesses/slug/{slug}:
    get:
      operationId: getBusinessBySlug
      tags: [businesses]
      summary: Get business by slug (public)
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Business"

  # ── Appointments ─────────────────────────────────────────────────────────────
...
  /appointments/{appointmentId}/slots:
    get:
      operationId: getAvailableSlots
      tags: [slots]
      parameters:
        - name: appointmentId
          in: path
          required: true
          schema:
            type: integer
        - name: date
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: providerId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TimeSlot"

  # ── Bookings ─────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        fullName:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [customer, organiser, admin]
        isActive:
          type: boolean
        isVerified:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, fullName, email, role, isActive, isVerified, createdAt]
...
    Business:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        category:
          type: string
        logoUrl:
          type: string
        coverUrl:
          type: string
        address:
          type: string
        city:
          type: string
        country:
          type: string
        phone:
          type: string
        website:
          type: string
        email:
          type: string
        ownerId:
          type: integer
        owner:
          $ref: "#/components/schemas/User"
        isActive:
          type: boolean
        isVerified:
          type: boolean
        appointmentCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, name, slug, category, ownerId, isActive, isVerified, createdAt]
    CreateBusinessBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        category:
          type: string
          enum: [healthcare, beauty_wellness, fitness, education, legal, financial, consulting, automotive, home_services, technology, hospitality, other]
        logoUrl:
          type: string
        coverUrl:
          type: string
        address:
          type: string
        city:
          type: string
        country:
          type: string
        phone:
          type: string
        website:
          type: string
        email:
          type: string
      required: [name, category]
    UpdateBusinessBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        category:
          type: string
          enum: [healthcare, beauty_wellness, fitness, education, legal, financial, consulting, automotive, home_services, technology, hospitality, other]
        logoUrl:
          type: string
        coverUrl:
          type: string
        address:
          type: string
        city:
          type: string
        country:
          type: string
        phone:
          type: string
        website:
          type: string
        email:
          type: string
        isActive:
          type: boolean
    BusinessMember:
      type: object
      properties:
        id:
          type: integer
        businessId:
          type: integer
        userId:
          type: integer
        user:
          $ref: "#/components/schemas/User"
        role:
          type: string
          enum: [owner, staff, admin]
        joinedAt:
          type: string
          format: date-time
      required: [id, businessId, userId, role, joinedAt]
    AddBusinessMemberBody:
      type: object
      properties:
        email:
          type: string
          format: email
        role:
          type: string
          enum: [owner, staff, admin]
      required: [email]
    BusinessStats:
      type: object
      properties:
        totalAppointmentTypes:
          type: integer
        publishedAppointmentTypes:
          type: integer
        totalBookings:
          type: integer
        confirmedBookings:
          type: integer
        pendingBookings:
          type: integer
        cancelledBookings:
          type: integer
        todayBookings:
          type: integer
        upcomingBookings:
          type: integer
        totalRevenue:
          type: number
        totalStaff:
          type: integer
      required: [totalAppointmentTypes, publishedAppointmentTypes, totalBookings, confirmedBookings, pendingBookings, cancelledBookings, todayBookings, upcomingBookings, totalRevenue, totalStaff]
...
    TimeSlot:
      type: object
      properties:
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        available:
          type: boolean
        remainingCapacity:
          type: integer
        providerId:
          type: integer
        providerName:
          type: string
      required: [startTime, endTime, available]
...